pyproject.toml
[tool.poetry]
name = "data-engineer-exam"
version = "0.1.0"
description = ""
authors = ["Ray Chen <ray941216@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.1"
matplotlib = "^3.1.1"
pandas = "^1.0.2"
torch = "*"
torchvision = "*"
ipython = "*"
click = "^7.0"
scikit-learn = "^0.23.1"
pendulum = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
